Reference

ParametricOptInterface.Optimizer — Type

Optimizer{T, OT <: MOI.ModelLike} <: MOI.AbstractOptimizer

Declares a Optimizer, which allows the handling of parameters in a optimization model.

Keyword arguments

evaluate_duals::Bool: If true, evaluates the dual of parameters. Users might want to set it to false to increase performance when the duals of parameters are not necessary. Defaults to true.

Example

julia> ParametricOptInterface.Optimizer(GLPK.Optimizer())
ParametricOptInterface.Optimizer{Float64,GLPK.Optimizer}

ParametricOptInterface.Parameter — Type

Parameter(val::Float64)

The Parameter structure stores the numerical value associated to a given parameter.

Example

julia> ParametricOptInterface.Parameter(5)
ParametricOptInterface.Parameter(5)

ParametricOptInterface.ParameterValue — Type

ParameterValue <: MOI.AbstractVariableAttribute

Attribute defined to set and get parameter values

Example

MOI.set(model, POI.ParameterValue(), p, 2.0)
MOI.get(model, POI.ParameterValue(), p)

ParametricOptInterface.ParameterDual — Type

ParameterDual <: MOI.AbstractVariableAttribute

Attribute defined to get the dual values associated to parameters

Example

MOI.get(model, POI.ParameterValue(), p)